requirements.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
requests>=2.31.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
pyarrow>=12.0.0
jupyterlab>=4.0.0
fredapi>=0.5.0

